Mv fcts out from simulations
- expand_model belongs to models
- estimate_gridding_options belongs to meshes
