+++ tests/test_meshes.py
@@ -8,12 +8,19 @@ from numpy.testing import assert_allclose
 import emg3d
 from emg3d import meshes
 
+from . import helpers
+
 
 # Import soft dependencies.
 try:
     import discretize
 except ImportError:
     discretize = None
+
+try:
+    import xarray
+except ImportError:
+    xarray = None
 
 # Data generated with create_data/regression.py
 REGRES = emg3d.load(join(dirname(__file__), 'data', 'regression.npz'))
@@ -461,3 +468,145 @@ def test_check_mesh():
     hx = np.ones(16)*20
     grid = meshes.TensorMesh(h=[hx, hx, hx], origin=(0, 0, 0))
     meshes.check_mesh(grid)
+
+
+@pytest.mark.skipif(xarray is None, reason="xarray not installed.")
+class TestEstimateGriddingOpts():
+    if xarray is not None:
+        # Create a simple survey
+        sources = emg3d.surveys.txrx_coordinates_to_dict(
+                emg3d.TxElectricDipole,
+                (0, [1000, 3000, 5000], -950, 0, 0))
+        receivers = emg3d.surveys.txrx_coordinates_to_dict(
+                emg3d.RxElectricPoint,
+                (np.arange(11)*500, 2000, -1000, 0, 0))
+        frequencies = (0.1, 10.0)
+
+        survey = emg3d.Survey(
+                sources, receivers, frequencies, noise_floor=1e-15,
+                relative_error=0.05)
+
+        # Create a simple grid and model
+        grid = meshes.TensorMesh(
+                [np.ones(32)*250, np.ones(16)*500, np.ones(16)*500],
+                np.array([-1250, -1250, -2250]))
+        model = emg3d.Model(grid, 0.1, np.ones(grid.shape_cells)*10)
+        model.property_y[5, 8, 3] = 100000  # Cell at source center
+
+    def test_empty_dict(self):
+        gdict = meshes.estimate_gridding_opts({}, self.model, self.survey)
+
+        assert gdict['frequency'] == 1.0
+        assert gdict['mapping'] == self.model.map.name
+        assert_allclose(gdict['center'], (0, 3000, -950))
+        assert_allclose(gdict['domain']['x'], (-500, 5500))
+        assert_allclose(gdict['domain']['y'], (600, 5400))
+        assert_allclose(gdict['domain']['z'], (-3651, -651))
+        assert_allclose(gdict['properties'], [100000, 10, 10, 10, 10, 10, 10])
+
+    def test_mapping_vector(self):
+        gridding_opts = {
+            'mapping': "LgConductivity",
+            'vector': 'xZ',
+            }
+        gdict = meshes.estimate_gridding_opts(
+                gridding_opts, self.model, self.survey)
+
+        assert_allclose(
+                gdict['properties'],
+                np.log10(1/np.array([100000, 10, 10, 10, 10, 10, 10])),
+                atol=1e-15)
+        assert_allclose(gdict['vector']['x'], self.grid.nodes_x)
+        assert gdict['vector']['y'] is None
+        assert_allclose(gdict['vector']['z'], self.grid.nodes_z)
+
+    def test_vector_domain_distance(self):
+        gridding_opts = {
+                'vector': 'Z',
+                'domain': (None, [-1000, 1000], None),
+                'distance': [[5, 10], None, None],
+                }
+        gdict = meshes.estimate_gridding_opts(
+                gridding_opts, self.model, self.survey)
+
+        assert gdict['vector']['x'] == gdict['vector']['y'] is None
+        assert_allclose(gdict['vector']['z'], self.model.grid.nodes_z)
+
+        assert gdict['domain']['x'] is None
+        assert gdict['domain']['y'] == [-1000, 1000]
+        assert gdict['domain']['z'] == [self.model.grid.nodes_z[0],
+                                        self.model.grid.nodes_z[-1]]
+        assert gdict['distance']['x'] == [5, 10]
+        assert gdict['distance']['y'] == gdict['distance']['z'] is None
+
+        # As dict
+        gridding_opts = {
+                'vector': 'Z',
+                'domain': {'x': None, 'y': [-1000, 1000], 'z': None},
+                'distance': {'x': [5, 10], 'y': None, 'z': None},
+                }
+        gdict = meshes.estimate_gridding_opts(
+                gridding_opts, self.model, self.survey)
+
+        assert gdict['vector']['x'] == gdict['vector']['y'] is None
+        assert_allclose(gdict['vector']['z'], self.model.grid.nodes_z)
+
+        assert gdict['domain']['x'] is None
+        assert gdict['domain']['y'] == [-1000, 1000]
+        assert gdict['domain']['z'] == [self.model.grid.nodes_z[0],
+                                        self.model.grid.nodes_z[-1]]
+        assert gdict['distance']['x'] == [5, 10]
+        assert gdict['distance']['y'] == gdict['distance']['z'] is None
+
+    def test_pass_along(self):
+        gridding_opts = {
+            'vector': {'x': None, 'y': 1, 'z': None},
+            'stretching': [1.2, 1.3],
+            'seasurface': -500,
+            'cell_numbers': [10, 20, 30],
+            'lambda_factor': 0.8,
+            'max_buffer': 10000,
+            'min_width_limits': ([20, 40], [20, 40], [20, 40]),
+            'min_width_pps': 4,
+            'verb': -1,
+            }
+
+        gdict = meshes.estimate_gridding_opts(
+                gridding_opts.copy(), self.model, self.survey)
+
+        # Check that all parameters passed unchanged.
+        gdict2 = {k: gdict[k] for k, _ in gridding_opts.items()}
+        # Except the tuple, which should be a dict now
+        gridding_opts['min_width_limits'] = {
+                'x': gridding_opts['min_width_limits'][0],
+                'y': gridding_opts['min_width_limits'][1],
+                'z': gridding_opts['min_width_limits'][2]
+        }
+        assert helpers.compare_dicts(gdict2, gridding_opts)
+
+    def test_factor(self):
+        sources = emg3d.TxElectricDipole((0, 3000, -950, 0, 0))
+        receivers = emg3d.RxElectricPoint((0, 3000, -1000, 0, 0))
+
+        # Adjusted x-domain.
+        survey = emg3d.Survey(
+                self.sources, receivers, self.frequencies, noise_floor=1e-15,
+                relative_error=0.05)
+
+        gdict = meshes.estimate_gridding_opts({}, self.model, survey)
+
+        assert_allclose(gdict['domain']['x'], (-800, 800))
+
+        # Adjusted x-domain.
+        survey = emg3d.Survey(
+                sources, self.receivers, self.frequencies, noise_floor=1e-15,
+                relative_error=0.05)
+
+        gdict = meshes.estimate_gridding_opts({}, self.model, survey)
+
+        assert_allclose(gdict['domain']['y'], (1500, 3500))
+
+    def test_error(self):
+        with pytest.raises(TypeError, match='Unexpected gridding_opts'):
+            _ = meshes.estimate_gridding_opts(
+                    {'what': True}, self.model, self.survey)
