+++ tests/test_models.py
@@ -412,3 +412,39 @@ class TestModelOperators:
         del mdict['grid']
         with pytest.raises(KeyError, match="'grid'"):
             models.Model.from_dict(mdict)
+
+
+def test_expand_grid_model():
+    grid = emg3d.TensorMesh([[4, 2, 2, 4], [2, 2, 2, 2], [1, 1]], (0, 0, 0))
+    model = emg3d.Model(grid, 1, np.ones(grid.shape_cells)*2, mu_r=3,
+                        epsilon_r=5)
+
+    o_model = models.expand_grid_model(model, [2, 3], 5)
+
+    # Grid.
+    assert_allclose(grid.nodes_z, o_model.grid.nodes_z[:-2])
+    assert o_model.grid.nodes_z[-2] == 5
+    assert o_model.grid.nodes_z[-1] == 105
+
+    # Property x (from float).
+    assert_allclose(o_model.property_x[:, :, :-2], 1)
+    assert_allclose(o_model.property_x[:, :, -2], 2)
+    assert_allclose(o_model.property_x[:, :, -1], 3)
+
+    # Property y (from shape_cells).
+    assert_allclose(o_model.property_y[:, :, :-2], model.property_y)
+    assert_allclose(o_model.property_y[:, :, -2], 2)
+    assert_allclose(o_model.property_y[:, :, -1], 3)
+
+    # Property z.
+    assert o_model.property_z is None
+
+    # Property mu_r (from float).
+    assert_allclose(o_model.mu_r[:, :, :-2], 3)
+    assert_allclose(o_model.mu_r[:, :, -2], 1)
+    assert_allclose(o_model.mu_r[:, :, -1], 1)
+
+    # Property epsilon_r (from float).
+    assert_allclose(o_model.epsilon_r[:, :, :-2], 5)
+    assert_allclose(o_model.epsilon_r[:, :, -2], 1)
+    assert_allclose(o_model.epsilon_r[:, :, -1], 1)
